Move inspector panel out of PostList
