+++ public_html/wp-content/mu-plugins/blocks/source/components/index.js
@@ -3,6 +3,7 @@
  */
 // Controls components
 export { default as EditAppender } from './edit-appender';
+export { default as GridInspectorPanel } from './grid-inspector-panel';
 
 // Item selector components
 export { buildOptions, getOptionLabel, ItemSelect, Option } from './item-select';
@@ -10,7 +11,6 @@ export { buildOptions, getOptionLabel, ItemSelect, Option } from './item-select'
 // Content components
 export { default as PostList } from './post-list';
 export { default as NoContent } from './post-list/no-content';
-export { default as GridInspectorPanel } from './post-list/inspector-controls';
 export { default as LayoutToolbar } from './post-list/toolbar';
 export { DangerousItemHTMLContent, ItemPermalink, ItemTitle } from './item';
